requirements.txt
bleach==3.1.0
bokeh==1.0.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.7
django-cron==0.5.1
fints==2.0.1
Flask==1.0.2
Flask-Compress==1.4.0
fuzzywuzzy==0.17.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mt-940==4.13.0
numpy==1.16.1
packaging==19.0
pandas==0.24.1
Pillow==5.4.1
psycopg2-binary==2.7.7
pyparsing==2.3.1
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2018.9
PyYAML==4.2b4
requests==2.21.0
sepaxml==2.0.0
six==1.12.0
SQLAlchemy==1.3.0b1
tornado==5.1.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
